Refactor index.ts
Refactored index.ts into separate modules in order to support unit testing:

- getArguments.ts
- getTestAssemblies.ts
- getVsTestPath.ts
- uploadArtifact.ts
